Feat/add output for user (#133)
* refactor: add error message for empty rows

* feat: add (statement completed) header when page token is empty

* refactor: proper string for completed

+++ internal/pkg/flink/pkg/controller/store.go
@@ -180,14 +180,17 @@ func (s *Store) FetchStatementResults(statement types.ProcessedStatement) (*type
 }
 
 func (s *Store) DeleteStatement(statementName string) {
-	httpResponse, err := s.client.DeleteStatement(context.Background(), statementName)
+	demoMode := s.appOptions != nil && s.appOptions.MOCK_STATEMENTS_OUTPUT_DEMO
+	if !demoMode {
+		httpResponse, err := s.client.DeleteStatement(context.Background(), statementName)
 
-	if err != nil {
-		log.Print(err.Error())
-	}
+		if err != nil {
+			log.Print(err.Error())
+		}
 
-	if httpResponse.StatusCode < 200 || httpResponse.StatusCode >= 300 {
-		log.Printf("Error: " + httpResponse.Body.Close().Error() + httpResponse.Status)
+		if httpResponse.StatusCode < 200 || httpResponse.StatusCode >= 300 {
+			log.Printf("Error: " + httpResponse.Body.Close().Error() + httpResponse.Status)
+		}
 	}
 }
